refactor(user-management): split type tests for GH727
